// src/types/places.ts
export interface CreatePlaceData {
  place_id: string;
  name?: string | null;
  formatted_address?: string | null;
  international_phone_number?: string | null;
  national_phone_number?: string | null;
  website?: string | null;
  url?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  viewport_ne_lat?: number | null;
  viewport_ne_lng?: number | null;
  viewport_sw_lat?: number | null;
  viewport_sw_lng?: number | null;
  business_status?: string | null;
  price_level?: string | null;
  rating?: number | null;
  user_ratings_total?: number | null;
  description?: string | null;
  open_now?: boolean | null;
  periods?: Record<string, any> | null;
  weekday_text?: string[] | null;
  photos?: Record<string, any> | null;
  types?: string[] | null;
  primary_type?: string | null;
  wheelchair_accessible_entrance?: boolean | null;
  delivery?: boolean | null;
  dine_in?: boolean | null;
  takeout?: boolean | null;
  reservable?: boolean | null;
  serves_beer?: boolean | null;
  serves_breakfast?: boolean | null;
  serves_brunch?: boolean | null;
  serves_dinner?: boolean | null;
  serves_lunch?: boolean | null;
  serves_vegetarian_food?: boolean | null;
  serves_wine?: boolean | null;
  atmosphere?: string | null;
  crowd_level?: string | null;
  curbside_pickup?: boolean | null;
  drive_through?: boolean | null;
  good_for_children?: boolean | null;
  good_for_groups?: boolean | null;
  outdoor_seating?: boolean | null;
  live_music?: boolean | null;
  parking?: boolean | null;
  parking_options?: Record<string, any> | null;
  payment_options?: Record<string, any> | null;
  price_range?: string | null;
  reviews?: Record<string, any> | null;
  review_summary?: string | null;
  formatted_phone_number?: string | null;
  icon?: string | null;
  icon_background_color?: string | null;
  icon_mask_base_uri?: string | null;
  maps_url?: string | null;
  place_id_short?: string | null;
  plus_code?: Record<string, any> | null;
  reference?: string | null;
  scope?: string | null;
  geometry?: Record<string, any> | null;
  address_components?: Record<string, any> | null;
  search_terms?: string[] | null;
}